+++ src/aks-preview/HISTORY.rst
@@ -11,7 +11,7 @@ To release a new version, please select a new version number (usually plus 1 to 
 
 Pending
 +++++++
-* Add interactive AI-powered debugging tool az aks agent.
+* Add framework for interactive AI-powered debugging tool.
 
 18.0.0b26
 +++++++